requirements.txt
Django==1.8.6 
MySQL-python==1.2.5
Unipath==1.1
django-inspectional-registration==0.6.0
django_migration_fixture==0.5.1
six==1.10.0
